application/flaskapp/requirements.txt
flask
SQLAlchemy
scikit-learn
boto3
python-dotenv